Built second layer of web page
Basic layout of second layer that contains
two containers. One on the left-side and one
on the right-side.

The left-side contains <p>, <sub>, <button>
tags.

The right-side contains an image.

From the first layer,
removed 100px padding from <nav> since it was not used.

Changed flex-direction to column of main container.